Resolve circular deps with toolkits
